+++ src/plugins/extra/backgrounds/Unsplash/constants.ts
@@ -1,8 +1,11 @@
+import { By } from './interfaces';
+
 export const UNSPLASH_API_KEY = process.env.UNSPLASH_API_KEY;
 export const UNSPLASH_UTM = '?utm_source=Start&utm_medium=referral&utm_campaign=api-credit';
 
 export const defaultProps = {
-  curated: true,
+  by: By.OFFICIAL,
+  collections: '',
   darken: true,
   featured: false,
   search: '',