+++ src/plugins/extra/backgrounds/Unsplash/interfaces.ts
@@ -1,5 +1,6 @@
 export interface Settings {
-  curated: boolean;
+  by: By;
+  collections: string;
   darken: boolean;
   featured: boolean;
   search: string;
@@ -11,4 +12,10 @@ export interface Image {
   location_title?: string;
   user_name: string;
   user_link: string;
+}
+
+export enum By {
+  OFFICIAL = 'official',
+  COLLECTIONS = 'collections',
+  SEARCH = 'search',
 }